Began process of splitting of all current files.
