+++ UMLAttribute.hpp
@@ -28,7 +28,6 @@ class UMLAttribute
 		string attributeName;
 
 	public:
-		// Params: string newAttribute (name of attribute)
 		// Constructor for attribute objects
 		UMLAttribute(string newAttribute);
 
@@ -39,21 +38,6 @@ class UMLAttribute
 		void changeName(string newAttributeName);
 
 };
-
-UMLAttribute::UMLAttribute(string newAttribute) 
-:attributeName(newAttribute)
-{
-}
-
-string UMLAttribute::getAttributeName() const
-{
-	return attributeName;
-}
-
-void UMLAttribute::changeName(string newAttributeName)
-{
-	attributeName = newAttributeName;
-}
 
 /************************************************************/
 #endif
